requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
gunicorn==20.0.4
idna==2.10
newsapi-python==0.2.6
psycopg2-binary==2.8.6
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
